# backend/docs/swagger.yaml
basePath: /api/v1
consumes:
- application/json
definitions:
...
  github_com_FNTR3455234_FinTrack_backend_internal_modelos.Cuenta:
    properties:
      archivada:
        type: boolean
      color:
        type: string
      id:
        type: string
      nombre:
        type: string
      saldo_inicial:
        type: number
      tipo:
        type: string
    type: object
...
  github_com_FNTR3455234_FinTrack_backend_internal_modelos.PeticionCuenta:
    properties:
      archivada:
        type: boolean
      color:
        type: string
      nombre:
        maxLength: 60
        minLength: 1
        type: string
      saldo_inicial:
        type: number
      tipo:
        enum:
        - efectivo
        - debito
        - credito
        - ahorro
        type: string
    required:
    - color
    - nombre
    - saldo_inicial
    - tipo
    type: object
...
  github_com_FNTR3455234_FinTrack_backend_internal_respuestas.Detalle:
    properties:
      codigo:
        type: string
      detalles:
        items:
          type: string
        type: array
      mensaje:
        type: string
    type: object
  github_com_FNTR3455234_FinTrack_backend_internal_respuestas.Meta:
    properties:
      limite:
        type: integer
      pagina:
        type: integer
      total:
        type: integer
      total_paginas:
        type: integer
    type: object
  github_com_FNTR3455234_FinTrack_backend_internal_respuestas.Sobre:
    properties:
      datos: {}
      meta:
        $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.Meta'
    type: object
  github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError:
    properties:
      error:
        $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.Detalle'
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: |-
    API REST de finanzas personales: cuentas, categorias, transacciones, presupuestos y reportes.
    Todos los datos pertenecen a un usuario. El identificador del dueño sale siempre del token, nunca del cuerpo de la peticion.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: FinTrack API
  version: 0.1.0
paths:
...
  /cuentas:
    get:
      description: Solo las del usuario del token.
      parameters:
      - description: Incluye tambien las archivadas
        in: query
        name: incluir_archivadas
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.Sobre'
            - properties:
                datos:
                  items:
                    $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_modelos.Cuenta'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError'
      security:
      - BearerAuth: []
      summary: Listar cuentas
      tags:
      - cuentas
    post:
      parameters:
      - description: Datos de la cuenta
        in: body
        name: cuerpo
        required: true
        schema:
          $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_modelos.PeticionCuenta'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.Sobre'
            - properties:
                datos:
                  $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_modelos.Cuenta'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError'
      security:
      - BearerAuth: []
      summary: Crear una cuenta
      tags:
      - cuentas
  /cuentas/{id}:
    delete:
      description: 'Responde 409 CUENTA_CON_TRANSACCIONES si tiene movimientos. No
        hay borrado en cascada: archivala en su lugar.'
      parameters:
      - description: Identificador de 24 caracteres
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Sin contenido
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError'
      security:
      - BearerAuth: []
      summary: Borrar una cuenta
      tags:
      - cuentas
    get:
      description: 'Responde 404 si el identificador es de otro usuario: un 403 confirmaria
        que existe.'
      parameters:
      - description: Identificador de 24 caracteres
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.Sobre'
            - properties:
                datos:
                  $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_modelos.Cuenta'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError'
      security:
      - BearerAuth: []
      summary: Obtener una cuenta
      tags:
      - cuentas
    put:
      parameters:
      - description: Identificador de 24 caracteres
        in: path
        name: id
        required: true
        type: string
      - description: Datos de la cuenta
        in: body
        name: cuerpo
        required: true
        schema:
          $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_modelos.PeticionCuenta'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.Sobre'
            - properties:
                datos:
                  $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_modelos.Cuenta'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError'
      security:
      - BearerAuth: []
      summary: Editar una cuenta
      tags:
      - cuentas
...
produces:
- application/json
securityDefinitions:
  BearerAuth:
    description: 'Token de acceso con el prefijo. Ejemplo: "Bearer eyJhbGciOi..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"